+++ src/models/options/DonutChartConfiguration.class.ts
@@ -2,8 +2,8 @@ import { PieChartConfiguration } from './PieChartConfiguration.class';
 
 export class DonutChartConfiguration extends PieChartConfiguration {
 
-  readonly type: string = 'donutChart';
-  readonly donut: boolean = true;
+  public readonly type: string = 'donutChart';
+  public readonly donut: boolean = true;
 
   public donutRatio: number = 0.35;